Refactor FreeScale.kt file and adjust import statements
