+++ kandy-api/src/main/kotlin/org/jetbrains/kotlinx/kandy/ir/scale/FreeScale.kt
@@ -2,9 +2,11 @@
 * Copyright 2020-2023 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
 */
 
-package org.jetbrains.kotlinx.kandy.ir.bindings
+package org.jetbrains.kotlinx.kandy.ir.scale
 
 import org.jetbrains.kotlinx.kandy.ir.aes.AesName
+import org.jetbrains.kotlinx.kandy.ir.bindings.MappingParameters
+import org.jetbrains.kotlinx.kandy.ir.bindings.PositionalMappingParameters
 
 /**
  * Defines constraints and conditions for an axis.
