+++ kandy-lets-plot/src/main/kotlin/org/jetbrains/kotlinx/kandy/letsplot/layers/context/aes/WithX.kt
@@ -7,8 +7,8 @@ package org.jetbrains.kotlinx.kandy.letsplot.layers.context.aes
 import org.jetbrains.kotlinx.dataframe.DataColumn
 import org.jetbrains.kotlinx.dataframe.columns.ColumnReference
 import org.jetbrains.kotlinx.kandy.dsl.internal.BindingContext
-import org.jetbrains.kotlinx.kandy.ir.bindings.PositionalFreeScale
 import org.jetbrains.kotlinx.kandy.ir.bindings.PositionalMapping
+import org.jetbrains.kotlinx.kandy.ir.scale.PositionalFreeScale
 import org.jetbrains.kotlinx.kandy.letsplot.internal.LetsPlotPositionalMappingParametersContinuous
 import org.jetbrains.kotlinx.kandy.letsplot.internal.X
 import org.jetbrains.kotlinx.kandy.letsplot.scales.guide.model.AxisParametersWithSetter
@@ -23,21 +23,33 @@ public interface WithX : BindingContext {
         column: ColumnReference<T>,
         parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
     ): PositionalMapping<T> {
-        return addPositionalMapping<T>(X, column.name(), LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters))
+        return addPositionalMapping<T>(
+            X,
+            column.name(),
+            LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+        )
     }
 
     public fun <T> x(
         column: KProperty<T>,
         parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
     ): PositionalMapping<T> {
-        return addPositionalMapping<T>(X, column.name, LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters))
+        return addPositionalMapping<T>(
+            X,
+            column.name,
+            LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+        )
     }
 
     public fun x(
         column: String,
         parameters: LetsPlotPositionalMappingParametersContinuous<Any?>.() -> Unit = {}
     ): PositionalMapping<Any?> {
-        return addPositionalMapping<Any?>(X, column, LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters))
+        return addPositionalMapping<Any?>(
+            X,
+            column,
+            LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters)
+        )
     }
 
     public fun <T> x(
