+++ kandy-lets-plot/src/main/kotlin/org/jetbrains/kotlinx/kandy/letsplot/xy.kt
@@ -7,8 +7,8 @@ package org.jetbrains.kotlinx.kandy.letsplot
 import org.jetbrains.kotlinx.dataframe.DataColumn
 import org.jetbrains.kotlinx.dataframe.columns.ColumnReference
 import org.jetbrains.kotlinx.kandy.dsl.internal.LayerPlotContext
-import org.jetbrains.kotlinx.kandy.ir.bindings.PositionalFreeScale
 import org.jetbrains.kotlinx.kandy.ir.bindings.PositionalMapping
+import org.jetbrains.kotlinx.kandy.ir.scale.PositionalFreeScale
 import org.jetbrains.kotlinx.kandy.letsplot.internal.LetsPlotPositionalMappingParametersContinuous
 import org.jetbrains.kotlinx.kandy.letsplot.internal.X
 import org.jetbrains.kotlinx.kandy.letsplot.internal.Y
@@ -16,15 +16,19 @@ import org.jetbrains.kotlinx.kandy.letsplot.scales.guide.model.AxisParametersWit
 
 public fun <T> LayerPlotContext.x(
     column: ColumnReference<T>,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
-    return addPositionalMapping<T>(X, column.name(), LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters))
+    return addPositionalMapping<T>(
+        X,
+        column.name(),
+        LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+    )
 }
 
 public fun <T> LayerPlotContext.x(
     values: Iterable<T>,
     name: String? = null,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
     return addPositionalMapping<T>(
         X,
@@ -36,20 +40,24 @@ public fun <T> LayerPlotContext.x(
 
 public fun LayerPlotContext.x(
     column: String,
-    parameters:LetsPlotPositionalMappingParametersContinuous<Any?>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<Any?>.() -> Unit = {}
 ): PositionalMapping<Any?> {
-    return addPositionalMapping<Any?>(X, column, LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters))
+    return addPositionalMapping<Any?>(
+        X,
+        column,
+        LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters)
+    )
 }
 
 public fun <T> LayerPlotContext.x(
     values: DataColumn<T>,
     //name: String? = null,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
     return addPositionalMapping<T>(
         X,
         values,
-       LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+        LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
     )
 }
 
@@ -57,7 +65,7 @@ public fun <T> LayerPlotContext.x(
 public val LayerPlotContext.x: AxisParametersWithSetter
     get() {
         return AxisParametersWithSetter(bindingCollector.freeScales.getOrPut(X) {
-            PositionalFreeScale(X,LetsPlotPositionalMappingParametersContinuous<Any?>())
+            PositionalFreeScale(X, LetsPlotPositionalMappingParametersContinuous<Any?>())
         }.parameters as LetsPlotPositionalMappingParametersContinuous<Any?>, X, this)
     }
 
@@ -76,40 +84,48 @@ public fun <T> LayerPlotContext.y(value: T): PositionalSetting<T> {
 
 public fun <T> LayerPlotContext.y(
     column: ColumnReference<T>,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
-    return addPositionalMapping<T>(Y, column.name(),LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters))
+    return addPositionalMapping<T>(
+        Y,
+        column.name(),
+        LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+    )
 }
 
 public fun <T> LayerPlotContext.y(
     values: Iterable<T>,
     name: String? = null,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
     return addPositionalMapping<T>(
         Y,
         values.toList(),
         name,
-       LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+        LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
     )
 }
 
 public fun LayerPlotContext.y(
     column: String,
-    parameters:LetsPlotPositionalMappingParametersContinuous<Any?>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<Any?>.() -> Unit = {}
 ): PositionalMapping<Any?> {
-    return addPositionalMapping<Any?>(Y, column,LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters))
+    return addPositionalMapping<Any?>(
+        Y,
+        column,
+        LetsPlotPositionalMappingParametersContinuous<Any?>().apply(parameters)
+    )
 }
 
 public fun <T> LayerPlotContext.y(
     values: DataColumn<T>,
     //name: String? = null,
-    parameters:LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
+    parameters: LetsPlotPositionalMappingParametersContinuous<T>.() -> Unit = {}
 ): PositionalMapping<T> {
     return addPositionalMapping<T>(
         Y,
         values,
-       LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
+        LetsPlotPositionalMappingParametersContinuous<T>().apply(parameters)
     )
 }
 
@@ -117,7 +133,7 @@ public fun <T> LayerPlotContext.y(
 public val LayerPlotContext.y: AxisParametersWithSetter
     get() {
         return AxisParametersWithSetter(bindingCollector.freeScales.getOrPut(Y) {
-            PositionalFreeScale(Y,LetsPlotPositionalMappingParametersContinuous<Any?>())
+            PositionalFreeScale(Y, LetsPlotPositionalMappingParametersContinuous<Any?>())
         }.parameters as LetsPlotPositionalMappingParametersContinuous<Any?>, Y, this)
     }
